buy fix
it works. but it is not very good. because there is a notably gap between two wav files when they are being played. Is it possible to use javascript to combine all these wav files into a actual single wav file(combined.wav), and then use the combined wav file to play in the audio tag in the new area